requirements.txt
pika==1.3.2
openai==1.6.1
python-dotenv==1.0.0
pymongo==4.6.1
flask==3.0.0